+++ include/md.hpp
@@ -36,6 +36,7 @@
 #include "md/forcefield/intra_subsystem_pair_forcefield.hpp"
 #include "md/forcefield/neighbor_pair_forcefield.hpp"
 #include "md/forcefield/neighbor_pair_forcefield_v2.hpp"
+#include "md/forcefield/plane_surface_forcefield.hpp"
 #include "md/forcefield/point_source_forcefield.hpp"
 #include "md/forcefield/selected_pair_forcefield.hpp"
 #include "md/forcefield/sequential_pair_forcefield.hpp"